# generated/openapi-specification.json
{"openapi":"3.0.1","info":{"title":"OpenAPI - Sample Project","description":"This service aims to provide information regarding Orders...","contact":{"name":"Fabricio Entringer","email":"moreira@gmal.com"},"version":"1.0"},"externalDocs":{"description":"User Requirements","url":"https://openapi-sample.org/docs"},"servers":[{"url":"http://localhost:8080"},{"url":"https://openapi-sample.org"}],"security":[{"Logged User":["ADMIM","USER"]}],"paths":{"/{id}":{"get":{"tags":["Find One"],"summary":"Find One Order","description":"This service returns one Order stored in the database that matches with the id informed by the parameter.","operationId":"findOrder","parameters":[{"name":"id","in":"path","description":"Order Unique Identification","required":true,"schema":{"type":"integer","format":"int32"},"example":1}],"responses":{"200":{"description":"Order found successful.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Order"}}}},"404":{"description":"Order not found."}},"security":[{"Users and Admins":["USER","ADMIN"]}]},"put":{"tags":["Edit Order"],"summary":"Edit Order","description":"This service edits an existing Order and return its values after stored in the database.","operationId":"editOrder","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Order"}}},"required":true},"responses":{"202":{"description":"Edition accepted.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Order"}}}},"400":{"description":"Bad Request."}},"security":[{"Users and Admins":["USER","ADMIN"]}]},"delete":{"tags":["Insert Order"],"summary":"Delete Order","description":"This service deletes an existing Order stored in the database.","operationId":"deleteOrder","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"202":{"description":"Deletion accepted."},"400":{"description":"Bad Request."}},"security":[{"Users and Admins":["USER","ADMIN"]}]}},"/":{"get":{"tags":["Find All"],"summary":"Find all Orders","description":"This service returns all Orders stored in the database.","operationId":"findAllOrders","responses":{"200":{"description":"Order found successful.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Order"}}}}},"204":{"description":"No data found."}},"security":[{"Users and Admins":["USER","ADMIN"]}]},"post":{"tags":["Insert New Order"],"summary":"Insert New Order","description":"This service inserts a new Order and return its values after stored in the database.","operationId":"insertOrder","requestBody":{"description":"Order to be persisted.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Order"}}},"required":true},"responses":{"201":{"description":"Order Created.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Order"}}}},"400":{"description":"Bad Request."}},"security":[{"Users and Admins":["USER","ADMIN"]}]}}},"components":{"schemas":{"Customer":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"}}},"Order":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"product":{"$ref":"#/components/schemas/Product"},"customer":{"$ref":"#/components/schemas/Customer"},"amount":{"type":"integer","format":"int32"},"total":{"type":"number"}}},"Product":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"description":{"type":"string"}}}},"securitySchemes":{"BearerJWT":{"type":"openIdConnect","description":"Authentication via Keycloak/OAUTH2","in":"header","bearerFormat":"JWT"}}}}